+++ app/src/main/java/ar/com/wolox/android/example/di/AppComponent.kt
@@ -2,7 +2,7 @@ package ar.com.wolox.android.example.di
 
 import android.app.Application
 import ar.com.wolox.android.example.BootstrapApplication
-import ar.com.wolox.android.example.ui.example.ExampleModule
+import ar.com.wolox.android.example.ui.login.LoginModule
 import ar.com.wolox.android.example.ui.viewpager.ViewPagerActivityModule
 import ar.com.wolox.wolmo.core.di.modules.ContextModule
 import ar.com.wolox.wolmo.core.di.modules.DefaultModule
@@ -16,7 +16,7 @@ import dagger.android.support.AndroidSupportInjectionModule
 @ApplicationScope
 @Component(dependencies = [NetworkingComponent::class],
         modules = [AndroidSupportInjectionModule::class, DefaultModule::class, ContextModule::class,
-            ExampleModule::class, ViewPagerActivityModule::class])
+            LoginModule::class, ViewPagerActivityModule::class])
 interface AppComponent : AndroidInjector<BootstrapApplication> {
 
     @Component.Builder
